Added comment to all functions using doxygen format

+++ inc/lib/memory.h
@@ -22,16 +22,37 @@ typedef void (*FREE_PTR)(void*);
 typedef void* (*MALLOC_PTR)(size_t);
 typedef void* (*MALLOC_ALIGNED_PTR)(size_t, size_t);
 
+// Copys n bytes from src to dst
+// @param dst The destination buffer
+// @param src The soruce buffer
+// @param n The number of bytes to copy
 void* memcpy(void* dst, const void* src, size_t n);
+
+// Sets n bytes of dst to value
+// @param dst The destination buffer
+// @param size The number of bytes to set
+// @param value The value to write
+// @return a pointer to the destination buffer
 void* memset(void* dst, size_t size, uint8_t value);
 
-inline void* void_ptr_offset_bytes(void* ptr, int offset);
+// Adds a offest to the ptr as if it were uint8_t*
+// @param ptr The pointer to offset
+// @param offset The number of bytes to add (can be negitive)
+// @return The pointer with the offset
+inline void* void_ptr_offset_bytes(void* ptr, ptrdiff_t offset);
 
+// Peformce a bitwise and on the pointer as if it is a number
+// @param ptr The pointer modify 
+// @param mask The mask to bitwise and with
+// @return The modifyed pointer
 inline void* void_ptr_bitwise_and(void* ptr, size_t mask);
 
+// Uses the tranlation table to the physical address of a pointer
+// @param virtual_ptr Pointer to translate
+// @return The translated pointer
 inline void* get_physical_address(void* virtual_ptr);
 
-inline void* void_ptr_offset_bytes(void* ptr, int offset)
+inline void* void_ptr_offset_bytes(void* ptr, ptrdiff_t offset)
 {
     return ((uint8_t*)ptr) + offset;
 }
